Refactored DAO layer, added Test & new Criteria API.

+++ CapitalBusConnectRepository/src/main/java/eu/sanjin/kurelic/cbc/repo/dao/BusLineDao.java
@@ -8,6 +8,6 @@ public interface BusLineDao {
 
     List<BusLine> getCityLines(int offset, int limit);
 
-    int getNumberOfCityLines();
+    Long getNumberOfCityLines();
 
 }
